+++ tests/testthat/test-shiny.R
@@ -24,6 +24,10 @@ test_that("shiny widgets works", {
 
   inst_deps()
 
+  if (!shinytest::dependenciesInstalled()) {
+    skip("shinytest - dependencies not installed")
+  }
+
   nowats <- getOption("AutoUnloadShiny")
   options(AutoUnloadShiny = FALSE)
 
@@ -74,7 +78,7 @@ test_that("shiny widgets works", {
 
 
   # if testing for images enable this
-  # also set when plotly present
+  # also set when plotly not present (screenshots are with plotly)
   # like test_image <- TRUE
   # else set test_image <- FALSE
   test_image <- FALSE